fix: hide form when success or error mndal appears
