+++ components/Landing/Popup/Popup.tsx
@@ -10,7 +10,7 @@ const Popup: React.FC<PopupTypes> = (props) => {
   return (
     <>
       <Backdrop />
-      <Modal className='flex flex-col items-center justify-center fixed top-[50%] left-[50%] translate-y-[-50%] translate-x-[-50%] bg-mainDark2 rounded-[10px]  z-40 w-[79%] md:w-[55%] lg:w-[45%] xl:w-[35%]'>
+      <Modal className='flex flex-col items-center justify-center fixed top-[50%] left-[50%] translate-y-[-50%] translate-x-[-50%] bg-profileDialog rounded-[10px]  z-40 w-[79%] md:w-[55%] lg:w-[45%] xl:w-[35%]'>
         <div className='flex flex-col justify-center items-center w-full gap-8 '>
           <div className='mt-6'>
             <Image src={icon} alt='popup-icon' width={70} height={70} />
