+++ components/Landing/PopupComponent/PopupComponent.tsx
@@ -1,16 +1,8 @@
-import { useTranslation } from 'react-i18next';
-import { useRouter } from 'next/router';
 import { Popup } from 'components';
-import { useContext } from 'react';
-import { AuthContext } from 'store';
+import { usePopupComponent } from './usePopupComponent';
 
 const PopupComponent = () => {
-  const router = useRouter();
-  const { t } = useTranslation();
-  const ctx = useContext(AuthContext);
-  const modalState = router.query.modal;
-  const passwordModalState = router.query.token;
-
+  const { t, modalState } = usePopupComponent();
   return (
     <>
       {modalState === 'account-verified' && (
@@ -22,7 +14,7 @@ const PopupComponent = () => {
           url={'/'}
         />
       )}
-      verified
+
       {modalState === 'email-verified' && (
         <Popup
           icon={'/assets/images/check-icon.png'}
@@ -60,13 +52,13 @@ const PopupComponent = () => {
           url={'https://gmail.com'}
         />
       )}
-      {passwordModalState && ctx.changePasswordUpdateState(true)}
+
       {modalState === 'password-updated-successfully' && (
         <Popup
           icon={'/assets/images/check-icon.png'}
           heading={t('home:passSuccH1')}
           paragraph={t('home:passSuccH2')}
-          buttonTxt={t('home:passBtn')}
+          buttonTxt={t('home:thanksBtn2')}
           url={'/'}
         />
       )}
