+++ public/locales/en/home.json
@@ -1,1 +1,1 @@
-{"commonRequired":"This field is required","confirmPassPlaceholder":"Confirm password","createAccount":"Create an account","emailPlaceholder":"Enter your email","emailRequired":"Email field is required","failed":"Activation Failed!","failedTxt":"There was a mistake, please try again","failedTxtBtn":"Go back to main page","forgotBack":"Back to log in","forgotBtn":"Send instructions","forgotH1":"Forgot password?","forgotPass":"Forgot password","forgotPopBack":"Skip, I’ll confirm later","forgotPopBtn":"Go to my email","forgotPopH1":"Check your email","forgotPopH2":"We have sent a password recover instructions \nto your email","fotgotH2":"Enter the email and we’ll send an email with\ninstructions to reset your password","haveAccount":"Already have an account?","inputConfirmPass":"Confirm password","inputEmail":"Email","inputName":"Name","inputPassword":"Password","invalidEmail":"Invalid email format","login":"Log in","logIn":"Log in","loginBtn":"Sign in","loginFooter":"Do not have an account?","loginFooter1":"Sign up","loginGoogle":"Log in with Google","loginH1":"Log in to your account","loginH2":"Welcome back! Please enter your details.","loginInp1":"Email","loginInp2":"Password","lowercase":"Values should be lowercase","MainH1":"Find any quote in millions of movie lines","max15":"max 15 characters","min3":"min 3 characters","min8":"min 8 characters","namePlaceholder":"At least 3 & max.15 lower case characters","newPassBtn":"Reset password","newPassH1":"Create new password","newPassH2":"Your new password must be different from \nprevious used passwords","numbers":"Enter numbers only","passBtn":"Go to my news feed","PassErrBtn":"Go to home page","passErrH1":"Failed","passErrH2":"Sorry, something went wrong, please try again","passRequired":"Password field is required","passSuccH1":"Success!","passSuccH2":"Your Password changed successfully","passwordDontMatch":"Password don't match","passwordPlaceholder":"At least 8 & max.15 lower case characters","quote1P":"Interstellar, 2014","quote1Qt":"You have to leave something behind to go forward","quote2P":"The Royal Tenenbaums, 2001","quote2Qt":"I think we’re just gonna have to be secretly in love with each other and leave it that","quote3P":"The lord of the rings, 2001","quote3Qt":"Not all those who wander are lost.","remember":"Rememer me","signup":"Sign Up","signupFooter":"Already have an account?","signupGoogle":"Sign up with Google","start":"Get started","startJourney":"Start your journey!","thanks":"Thank you!","thanksBtn1":"Go to my email","thanksBtn2":"Go to my main page","thanksText1":"Please check your email and follow the instructions to\nactivate your account.","thanksText2":"Your account has been activated.","user":"User","userExists":"User already exists","usernameRequired":"Username field is required","userPlaceholder":"Enter your email/username","userRequired":"User field is required","validPass":"Enter valid password","validUsername":"Enter valid username"}
+{"commonRequired":"This field is required","confirmPassPlaceholder":"Confirm password","createAccount":"Create an account","emailPlaceholder":"Enter your email","emailRequired":"Email field is required","failed":"Activation Failed!","failedTxt":"There was a mistake, please try again","failedTxtBtn":"Go back to main page","forgotBack":"Back to log in","forgotBtn":"Send instructions","forgotH1":"Forgot password?","forgotPass":"Forgot password","forgotPopBack":"Skip, I’ll confirm later","forgotPopBtn":"Go to my email","forgotPopH1":"Check your email","forgotPopH2":"We have sent a password recover instructions \nto your email","fotgotH2":"Enter the email and we’ll send an email with\ninstructions to reset your password","haveAccount":"Already have an account?","inputConfirmPass":"Confirm password","inputEmail":"Email","inputName":"Name","inputPassword":"Password","invalidEmail":"Invalid email format","login":"Log in","logIn":"Log in","loginBtn":"Sign in","loginFooter":"Do not have an account?","loginFooter1":"Sign up","loginGoogle":"Log in with Google","loginH1":"Log in to your account","loginH2":"Welcome back! Please enter your details.","loginInp1":"Email","loginInp2":"Password","lowercase":"Values should be lowercase","MainH1":"Find any quote in millions of movie lines","max15":"max 15 characters","min3":"min 3 characters","min8":"min 8 characters","namePlaceholder":"At least 3 & max.15 lower case characters","newPassBtn":"Reset password","newPassH1":"Create new password","newPassH2":"Your new password must be different from \nprevious used passwords","numbers":"Enter numbers only","passBtn":"Go to my news feed","PassErrBtn":"Go to home page","passErrH1":"Failed","passErrH2":"Sorry, something went wrong, please try again","passRequired":"Password field is required","passSuccH1":"Success!","passSuccH2":"Your Password changed successfully","passwordDontMatch":"Password don't match","passwordPlaceholder":"At least 8 & max.15 lower case characters","quote1P":"Interstellar, 2014","quote1Qt":"You have to leave something behind to go forward","quote2P":"The Royal Tenenbaums, 2001","quote2Qt":"I think we’re just gonna have to be secretly in love with each other and leave it that","quote3P":"The lord of the rings, 2001","quote3Qt":"Not all those who wander are lost.","remember":"Rememer me","signup":"Sign Up","signupFooter":"Already have an account?","signupGoogle":"Sign up with Google","start":"Get started","startJourney":"Start your journey!","thanks":"Thank you!","thanksBtn1":"Go to my email","thanksBtn2":"Go to my main page","thanksText1":"Please check your email and follow the instructions to\nactivate your account.","thanksText2":"Your account has been activated.","user":"User","userExists":"User already exists","usernameRequired":"Username field is required","userPlaceholder":"Enter your email/username","userRequired":"User field is required","validPass":"Enter valid password","validUsername":"Enter valid username","verified":"Verified","verifiedText":"Your new email is verified"}
