+++ public/locales/ge/home.json
@@ -1,1 +1,1 @@
-{"commonRequired":"სავალდებულო ველი","confirmPassPlaceholder":"გაიმეორე პაროლი","createAccount":"პროფილის შექმნა","emailPlaceholder":"ელფოსტა","emailRequired":"აუცილებელი ველი","failed":"გააქტიურება ვერ მოხერხდა!","failedTxt":"მოხდა შეცდომა, გთხოვთ სცადოთ ხელახლა","failedTxtBtn":"დაუბრუნდით მთავარ გვერდზე","forgotBack":"უკან","forgotBtn":"ინსტრუქციების გაგზავნა","forgotH1":"დაგავიწყდა პაროლი?","forgotPass":"პაროლის აღდგენა","forgotPopBack":"მოგვიანებით იყოს","forgotPopBtn":"მეილზე გადასვლა","forgotPopH1":"გადაამოწმე ელფოსტა","forgotPopH2":"ჩვენ გამოგიგზავნეთ პაროლის აღდგენის ინსტრუქციები შენს ელფოსტაზე","fotgotH2":"შეიყვანე ელფოსტა, ჩვენ გამოგიგზავნით ელფოსტას შენი პაროლის აღდგენის ინსტრუქციებით","haveAccount":"გაქვს ანგარიში?","inputConfirmPass":"გაიმეორე პაროლი","inputEmail":"ელ-ფოსტა","inputName":"სახელი","inputPassword":"პაროლი","invalidEmail":"ელფოსტის ფორმატი არასწორია","login":"ავტორიზაცია","logIn":"ავტორიზაცია","loginBtn":"ავტორიზაცია","loginFooter":"არ გაქვთ ანგარიში?","loginFooter1":"რეგისტრაცია","loginGoogle":"Google ავტორიზაცია","loginH1":"ავტორიზაცია","loginH2":"კეთილი იყოს შენი დაბრუნება! შეიყვანე შენი მონაცემები","loginInp1":"ელფოსტა","loginInp2":"პაროლი","lowercase":"გამოიყენე დაბალი რეგისტრი","MainH1":"იპოვე ნებისმიერი ციტატა მილიონობით ფილმის სტრიქონში","max15":"მაქს. 15 სიმბოლო","min3":"მინ. 3 სიმბოლო","min8":"მინ. 8 სიმბოლო","namePlaceholder":"მინიმუმ 3 და მაქსიმუმ 15 პატარა ასო","newPassBtn":"პაროლის დარესეტება","newPassH1":"ახალი პაროლი","newPassH2":"შენი ახალი პაროლი უნდა განსხვავდებოდეს ძველი პაროლებისგან","numbers":"შეიყვანეთ მხოლოდ ნომრები","passBtn":"გადადი ნიუს ფიდზე","PassErrBtn":"Გადადი მთავარ გვერდზე გვერდი","passErrH1":"მოხდა შეცდომა","passErrH2":"უკაცრავად, რაღაც პრობლემაა, გთხოვ, სცადო თავიდან","passRequired":"აუცილებელი ველი","passSuccH1":"მადლობა!","passSuccH2":"შენი პაროლი წარმატებით შეიცვალა","passwordDontMatch":"პაროლები არ ემთხვევა","passwordPlaceholder":"პაროლი","quote1P":"ვარსკვლავთშორისი, 2014 წ","quote1Qt":"წინ წასასვლელად რაღაც უნდა დატოვო","quote2P":"სამეფო ტენენბაუმები, 2001 წ","quote2Qt":"ვფიქრობ, ჩვენ უბრალოდ ფარულად უნდა შევიყვაროთ ერთმანეთი და ასე დავტოვოთ","quote3P":"ბეჭდების მბრძანებელი, 2001 წ","quote3Qt":"ყველა მოხეტიალე არ არის დაკარგული.","remember":"დამიმახსოვრე","signup":"რეგისტრაცია","signupFooter":"უკვე გაქვს ანგარიში?","signupGoogle":"Google რეგისტრაცია","start":"დაიწყე","startJourney":"დაიწყე შენი მოგზაურობა!","thanks":"Გმადლობთ!","thanksBtn1":"გადადით ჩემს მეილზე","thanksBtn2":"გადადი მთავარ გვერდზე","thanksText1":"შეამოწმე ელ-ფოსტა და მიყევი ინსტრუქციას","thanksText2":"შენი ანგარიში გააქტიურდა.","user":"მომხმარებელი","userExists":"Მომხმარებელი უკვე არსებობს","usernameRequired":"აუცილებელი ველი","userPlaceholder":"შეიყვანეთ შენი ელფოსტა/იუზერი","userRequired":"აუცილებელი ველი","validPass":"შეიყვანე ვალიდური პაროლი","validUsername":"შეიყვანეთ ვალიდური ნიკნეიმი"}
+{"commonRequired":"სავალდებულო ველი","confirmPassPlaceholder":"გაიმეორე პაროლი","createAccount":"პროფილის შექმნა","emailPlaceholder":"ელფოსტა","emailRequired":"აუცილებელი ველი","failed":"გააქტიურება ვერ მოხერხდა!","failedTxt":"მოხდა შეცდომა, გთხოვთ სცადოთ ხელახლა","failedTxtBtn":"დაუბრუნდით მთავარ გვერდზე","forgotBack":"უკან","forgotBtn":"ინსტრუქციების გაგზავნა","forgotH1":"დაგავიწყდა პაროლი?","forgotPass":"პაროლის აღდგენა","forgotPopBack":"მოგვიანებით იყოს","forgotPopBtn":"მეილზე გადასვლა","forgotPopH1":"გადაამოწმე ელფოსტა","forgotPopH2":"ჩვენ გამოგიგზავნეთ პაროლის აღდგენის ინსტრუქციები შენს ელფოსტაზე","fotgotH2":"შეიყვანე ელფოსტა, ჩვენ გამოგიგზავნით ელფოსტას შენი პაროლის აღდგენის ინსტრუქციებით","haveAccount":"გაქვს ანგარიში?","inputConfirmPass":"გაიმეორე პაროლი","inputEmail":"ელ-ფოსტა","inputName":"სახელი","inputPassword":"პაროლი","invalidEmail":"ელფოსტის ფორმატი არასწორია","login":"ავტორიზაცია","logIn":"ავტორიზაცია","loginBtn":"ავტორიზაცია","loginFooter":"არ გაქვთ ანგარიში?","loginFooter1":"რეგისტრაცია","loginGoogle":"Google ავტორიზაცია","loginH1":"ავტორიზაცია","loginH2":"კეთილი იყოს შენი დაბრუნება! შეიყვანე შენი მონაცემები","loginInp1":"ელფოსტა","loginInp2":"პაროლი","lowercase":"გამოიყენე დაბალი რეგისტრი","MainH1":"იპოვე ნებისმიერი ციტატა მილიონობით ფილმის სტრიქონში","max15":"მაქს. 15 სიმბოლო","min3":"მინ. 3 სიმბოლო","min8":"მინ. 8 სიმბოლო","namePlaceholder":"მინიმუმ 3 და მაქსიმუმ 15 პატარა ასო","newPassBtn":"პაროლის დარესეტება","newPassH1":"ახალი პაროლი","newPassH2":"შენი ახალი პაროლი უნდა განსხვავდებოდეს ძველი პაროლებისგან","numbers":"შეიყვანეთ მხოლოდ ნომრები","passBtn":"გადადი ნიუს ფიდზე","PassErrBtn":"Გადადი მთავარ გვერდზე გვერდი","passErrH1":"მოხდა შეცდომა","passErrH2":"უკაცრავად, რაღაც პრობლემაა, გთხოვ, სცადო თავიდან","passRequired":"აუცილებელი ველი","passSuccH1":"მადლობა!","passSuccH2":"შენი პაროლი წარმატებით შეიცვალა","passwordDontMatch":"პაროლები არ ემთხვევა","passwordPlaceholder":"პაროლი","quote1P":"ვარსკვლავთშორისი, 2014 წ","quote1Qt":"წინ წასასვლელად რაღაც უნდა დატოვო","quote2P":"სამეფო ტენენბაუმები, 2001 წ","quote2Qt":"ვფიქრობ, ჩვენ უბრალოდ ფარულად უნდა შევიყვაროთ ერთმანეთი და ასე დავტოვოთ","quote3P":"ბეჭდების მბრძანებელი, 2001 წ","quote3Qt":"ყველა მოხეტიალე არ არის დაკარგული.","remember":"დამიმახსოვრე","signup":"რეგისტრაცია","signupFooter":"უკვე გაქვს ანგარიში?","signupGoogle":"Google რეგისტრაცია","start":"დაიწყე","startJourney":"დაიწყე შენი მოგზაურობა!","thanks":"Გმადლობთ!","thanksBtn1":"გადადით ჩემს მეილზე","thanksBtn2":"გადადი მთავარ გვერდზე","thanksText1":"შეამოწმე ელ-ფოსტა და მიყევი ინსტრუქციას","thanksText2":"შენი ანგარიში გააქტიურდა.","user":"მომხმარებელი","userExists":"Მომხმარებელი უკვე არსებობს","usernameRequired":"აუცილებელი ველი","userPlaceholder":"შეიყვანეთ შენი ელფოსტა/იუზერი","userRequired":"აუცილებელი ველი","validPass":"შეიყვანე ვალიდური პაროლი","validUsername":"შეიყვანეთ ვალიდური ნიკნეიმი","verified":"ვერიფიციცირებულია","verifiedText":"თქვენი ახალი ელფოსტა დადასტურებულია"}
