+++ tailwind.config.js
@@ -53,7 +53,7 @@ module.exports = {
         feedBackground:
           'linear-gradient(187.16deg, #181623 0.07%, #191725 51.65%, #0D0B14 98.75%)',
         profileDialog:
-          'linear-gradient(-50deg, rgba(239, 239, 239, 0.125) -1.81%, rgba(239, 239, 239, 0.00514528) 102.5%, rgba(1, 1, 1, 0.00260417) 102.51%, rgba(239, 239, 239, 0.95) 102.52%)',
+          'linear-gradient(-50deg, rgba(239, 239, 239, 0.055) -1.81%, rgba(239, 239, 239, 0.00514528) 102.5%, rgba(1, 1, 1, 0.00260417) 102.51%, rgba(239, 239, 239, 0.95) 102.52%)',
         image: "url('/assets/images/profile.png')",
       },
     },
